+++ addons/confirma/src/classes/TestingClass.cs
@@ -8,16 +8,16 @@ using Confirma.Types;
 
 namespace Confirma.Classes;
 
-public class TestClass
+public class TestingClass
 {
-	public Type Type { get; }
-	public bool IsParallelizable { get; }
-	public IEnumerable<TestMethod> TestMethods { get; }
+	public readonly Type Type;
+	public readonly bool IsParallelizable;
+	public readonly IEnumerable<TestingMethod> TestMethods;
 
 	private TestsProps _props;
 	private readonly Dictionary<string, LifecycleMethodData> _lifecycleMethods = new();
 
-	public TestClass(Type type)
+	public TestingClass(Type type)
 	{
 		Type = type;
 		TestMethods = TestDiscovery.DiscoverTestMethods(type);