+++ addons/confirma/src/helpers/Reflection.cs
@@ -9,23 +9,23 @@ namespace Confirma.Helpers;
 
 public static class Reflection
 {
-	public static Type[] GetClassesFromAssembly(Assembly assembly)
+	public static IEnumerable<Type> GetClassesFromAssembly(Assembly assembly)
 	{
-		return assembly.GetTypes().Where(type => type.IsClass && !type.IsAbstract).ToArray();
+		return assembly.GetTypes().Where(type => type.IsClass && !type.IsAbstract);
 	}
 
-	public static Type[] GetTestClassesFromAssembly(Assembly assembly)
+	public static IEnumerable<Type> GetTestClassesFromAssembly(Assembly assembly)
 	{
 		var types = assembly.GetTypes();
-		return types.Where(type => type.HasAttribute<TestClassAttribute>()).ToArray();
+		return types.Where(type => type.HasAttribute<TestClassAttribute>());
 	}
 
-	public static IEnumerable<TestClass> GetParallelTestClasses(IEnumerable<TestClass> classes)
+	public static IEnumerable<TestingClass> GetParallelTestClasses(IEnumerable<TestingClass> classes)
 	{
 		return classes.Where(tc => tc.IsParallelizable);
 	}
 
-	public static IEnumerable<TestClass> GetSequentialTestClasses(IEnumerable<TestClass> classes)
+	public static IEnumerable<TestingClass> GetSequentialTestClasses(IEnumerable<TestingClass> classes)
 	{
 		return classes.Where(tc => !tc.IsParallelizable);
 	}
@@ -38,11 +38,11 @@ public static class Reflection
 		));
 	}
 
-	public static MethodInfo[] GetTestMethodsFromType(Type type)
+	public static IEnumerable<MethodInfo> GetTestMethodsFromType(Type type)
 	{
 		return type.GetMethods().Where(method => method.CustomAttributes.Any(
 			attribute => attribute.AttributeType == typeof(TestCaseAttribute)
-		)).ToArray();
+		));
 	}
 
 	public static bool HasAttribute<T>(this object obj) where T : Attribute